reverse last changes, remove footer move shopping cart back
