+++ pages/CartOverview/index.js
@@ -1,0 +1,11 @@
+//import internal resources
+import Link from "next/link";
+
+export default function CartOverviewPage() {
+  return (
+    <>
+      <Link href={"/"}>back</Link>
+      <h1>Hello this is the overpage</h1>
+    </>
+  );
+}
